wip mod ip with array
