+++ src/eth.rs
@@ -1,9 +1,9 @@
 use core::mem;
 
-use crate::ne::NetEndian;
+use crate::types::U16;
 
 /// Ethernet header, which is present at the beginning of every Ethernet frame.
-#[repr(C)]
+#[repr(C, packed)]
 #[derive(Debug, Copy, Clone, Default)]
 #[cfg_attr(feature = "serde", derive(::serde::Serialize, ::serde::Deserialize))]
 pub struct EthHdr {
@@ -12,7 +12,7 @@ pub struct EthHdr {
     /// Source MAC address.
     pub src_addr: [u8; 6],
     /// Protocol which is encapsulated in the payload of the frame.
-    pub ether_type: NetEndian<u16>,
+    pub ether_type: U16,
 }
 
 impl EthHdr {
@@ -40,7 +40,7 @@ mod test {
             mem::transmute::<[u8; EthHdr::LEN], _>(data_stream.try_into().unwrap())
         };
 
-        assert_eq!(ethhdr.ether_type.get(), EtherType::Ipv4 as u16);
+        assert_eq!(ethhdr.ether_type.to_bits(), EtherType::Ipv4 as u16);
         assert_eq!(ethhdr.dst_addr, [0xFF_u8, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF]);
         assert_eq!(ethhdr.src_addr, [0x00, 0x11, 0x22, 0x33, 0x44, 0x55]);
     }
